# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /remote-work:
    get:
      operationId: listRemoteWorkRequests
      tags: [remote-work]
      summary: List remote work requests (admin sees all, employee sees own)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Remote work requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RemoteWorkRequest"
    post:
      operationId: createRemoteWorkRequest
      tags: [remote-work]
      summary: Submit a remote work request (employee only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRemoteWorkRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteWorkRequest"
  /remote-work/{id}:
    patch:
      operationId: updateRemoteWork
      tags: [remote-work]
      summary: Update a pending remote work request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRemoteWorkRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteWorkRequest"
    delete:
      operationId: deleteRemoteWork
      tags: [remote-work]
      summary: Cancel a pending remote work request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /remote-work/{id}/approve:
    post:
      operationId: approveRemoteWork
      tags: [remote-work]
      summary: Approve a remote work request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteWorkRequest"
  /remote-work/{id}/reject:
    post:
      operationId: rejectRemoteWork
      tags: [remote-work]
      summary: Reject a remote work request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteWorkRequest"
...
  /employees/{id}/salary-components:
    get:
      operationId: listSalaryComponents
      tags: [salary]
      summary: List salary components for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalaryComponent"
    post:
      operationId: createSalaryComponent
      tags: [salary]
      summary: Add a salary component (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalaryComponentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalaryComponent"
  /employees/{id}/salary-components/{componentId}:
    patch:
      operationId: updateSalaryComponent
      tags: [salary]
      summary: Update a salary component (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: componentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalaryComponentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalaryComponent"
    delete:
      operationId: deleteSalaryComponent
      tags: [salary]
      summary: Delete a salary component (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: componentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...
    SalaryComponent:
      type: object
      required: [id, employeeId, label, kind, valueType, value, isDeduction, isTaxable]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        label: { type: string }
        kind:
          type: string
          enum: [designation, commission, allowance, provident_fund, other]
        valueType:
          type: string
          enum: [fixed, percentage]
        value: { type: number }
        isDeduction: { type: boolean }
        isTaxable: { type: boolean }
        sortOrder: { type: integer }
    CreateSalaryComponentRequest:
      type: object
      required: [label, kind, valueType, value]
      properties:
        label: { type: string }
        kind:
          type: string
          enum: [designation, commission, allowance, provident_fund, other]
        valueType:
          type: string
          enum: [fixed, percentage]
        value: { type: number }
        isDeduction: { type: boolean }
        isTaxable: { type: boolean }
    UpdateSalaryComponentRequest:
      type: object
      required: [label, kind, valueType, value]
      properties:
        label: { type: string }
        kind:
          type: string
          enum: [designation, commission, allowance, provident_fund, other]
        valueType:
          type: string
          enum: [fixed, percentage]
        value: { type: number }
        isDeduction: { type: boolean }
        isTaxable: { type: boolean }
...
    AttachmentItem:
      type: object
      required: [url, name]
      properties:
        url: { type: string }
        name: { type: string }
    MentionedEmployee:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...
    RemoteWorkRequest:
      type: object
      required: [id, employeeId, date, status, reason, appliedAt, mentionedEmployeeIds, attachments]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        employeeName: { type: string }
        date: { type: string, format: date }
        reason: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachmentName: { type: string, nullable: true }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentItem"
        mentionedEmployeeIds:
          type: array
          items: { type: integer }
        mentionedEmployees:
          type: array
          items:
            $ref: "#/components/schemas/MentionedEmployee"
        status: { type: string, enum: [pending, approved, rejected] }
        appliedAt: { type: string, description: "ISO datetime" }
        reviewedAt: { type: string, nullable: true }
    CreateRemoteWorkRequest:
      type: object
      required: [date, reason]
      properties:
        date: { type: string, format: date }
        reason: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachmentName: { type: string, nullable: true }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentItem"
        mentionedEmployeeIds:
          type: array
          items: { type: integer }
    UpdateRemoteWorkRequest:
      type: object
      properties:
        date: { type: string, format: date }
        reason: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachmentName: { type: string, nullable: true }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentItem"
        mentionedEmployeeIds:
          type: array
          items: { type: integer }
...